+++ app/Http/Livewire/Admin/User/Role/Index.php
@@ -53,7 +53,10 @@ class Index extends Component
     }
     public function delete(Role $role)
     {
-        //$this->authorize('delete', $user);
+        if(!auth()->user()->can('admin_roles_delete')) {
+            return abort(403);
+        }
+
         $this->confirm(__('bap.are_you_sure'), [
             'toast' => false,
             'position' => 'center',
@@ -67,7 +70,10 @@ class Index extends Component
 
     public function confirmedDelete()
     {
-        //$this->authorize('delete', $this->watcher);
+        if(!auth()->user()->can('admin_roles_delete')) {
+            return abort(403);
+        }
+
         $this->role->delete();
         $this->emit('updateList');
         $this->alert(
@@ -78,7 +84,6 @@ class Index extends Component
 
     public function cancelledDelete()
     {
-        //$this->authorize('delete', $this->user);
         $this->alert(
             'success',
             __('bap.cancelled')
@@ -87,6 +92,10 @@ class Index extends Component
 
     public function mount()
     {
+        if(!auth()->user()->can('admin_roles_index')) {
+            return abort(403);
+        }
+
         $this->search = request()->query('search', $this->search);
     }
 
@@ -122,6 +131,10 @@ class Index extends Component
 
     public function deleteSelectedQuery()
     {
+        if(!auth()->user()->can('admin_roles_delete')) {
+            return abort(403);
+        }
+
         Role::query()
             ->whereIn('id', $this->selectedRoles)
             ->delete();
@@ -137,6 +150,10 @@ class Index extends Component
 
     public function render()
     {
+        if(!auth()->user()->can('admin_roles_index')) {
+            return abort(403);
+        }
+
         $roles = Role::where('name', 'LIKE', '%' . $this->search . '%')->orderBy($this->sortColumn, $this->sortDirection)->paginate($this->perPage);
         return view('livewire.admin.user.role.index', compact('roles'))->layout('layouts.admin');
     }
